requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
babel==2.16.0
chardet==5.2.0
convertdate==2.4.0
jdatetime>=3.10
Django==5.1.3
django-tenants==3.7.0
docxcompose==1.4.0
docxtpl==0.19.0
Jinja2==3.1.4
lxml==5.3.0
jdatetime
django-tenants
MarkupSafe==3.0.2
pillow==11.0.0
psycopg2==2.9.10
PyMeeus==0.5.12
python-bidi==0.6.6
python-docx==1.1.2
reportlab==4.2.5
six==1.17.0
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
